requirements.txt
python-dotenv
openai
langchain 
langchain_community 
langchain_openai 
openai 
neo4j 
unstructured 
python-magic-bin 
tiktoken